[user-service] refactor: simplify protocol structure
