--- user/src/application/protocol/create-new-user.protocol.ts
@@ -1,3 +1,0 @@
-export interface CreateUserRepository {
-  create: (user: any) => Promise<void>
-}
