--- user/src/application/protocol/delete-refresh-token-by-id.protocol.ts
@@ -1,3 +1,0 @@
-export interface DeleteRefreshTokenByIdRepository {
-  delete: (id: string) => Promise<void>
-}
