--- user/src/application/protocol/find-user-by-email.protocol.ts
@@ -1,5 +1,0 @@
-import { type User } from '@/domain/entity/user.entity'
-
-export interface FindUserByEmailRepository {
-  findByEmail: (email: string) => Promise<User | null>
-}
