--- user/src/application/protocol/get-refresh-token-by-token-and-user-id.protocol.ts
@@ -1,5 +1,0 @@
-import { type RefreshToken } from './save-refresh-token.protocol'
-
-export interface GetRefreshTokenByTokenAndUserIdRepository {
-  getByTokenAndUserId: (token: string, userId: string) => Promise<RefreshToken | null>
-}
