--- user/src/application/protocol/hash-password.protocol.ts
@@ -1,3 +1,0 @@
-export interface HashPassword {
-  hash: (password: string) => Promise<string>
-}
