+++ user/src/application/protocol/index.ts
@@ -1,12 +1,3 @@
-export * from './find-user-by-email.protocol'
-export * from './hash-password.protocol'
-export * from './create-new-user.protocol'
-export * from './password-validator.protocol'
-export * from './token-generator.protocol'
-export * from './message-sender.protocol'
-export * from './send-email-notification.protocol'
-export * from './save-refresh-token.protocol'
-export * from './user-exists.protocol'
-export * from './get-refresh-token-by-token-and-user-id.protocol'
-export * from './delete-refresh-token-by-id.protocol'
-export * from './token-validator.protocol'
+export * from './repository'
+export * from './util'
+export * from './service'
