+++ user/src/application/protocol/repository/user-repository.protocol.ts
@@ -1,0 +1,13 @@
+import { type User } from '@/domain/entity'
+
+export interface CreateUserRepository {
+  create: (user: any) => Promise<void>
+}
+
+export interface FindUserByEmailRepository {
+  findByEmail: (email: string) => Promise<User | null>
+}
+
+export interface UserExistsRepository {
+  exists: (userId: string) => Promise<boolean>
+}
