--- user/src/application/protocol/save-refresh-token.protocol.ts
@@ -1,9 +1,0 @@
-export interface RefreshToken {
-  id: string
-  token: string
-  userId: string
-}
-
-export interface SaveRefreshTokenRepository {
-  save: (data: RefreshToken) => Promise<void>
-}
