+++ user/src/application/protocol/service/index.ts
@@ -1,0 +1,2 @@
+export * from './mailing.protocol'
+export * from './messaging.protocol'
