--- user/src/application/protocol/token-generator.protocol.ts
@@ -1,3 +1,0 @@
-export interface TokenGenerator {
-  generate: (payload: any) => Promise<string>
-}
