--- user/src/application/protocol/user-exists.protocol.ts
@@ -1,3 +1,0 @@
-export interface UserExistsRepository {
-  exists: (userId: string) => Promise<boolean>
-}
