+++ user/src/application/protocol/util/auth-token.protocol.ts
@@ -7,3 +7,7 @@ export interface TokenValidatorResponse {
 export interface TokenValidator {
   validate: (token: string) => Promise<TokenValidatorResponse>
 }
+
+export interface TokenGenerator {
+  generate: (payload: any) => Promise<string>
+}
