+++ user/src/application/protocol/util/password-handling.protocol.ts
@@ -1,3 +1,7 @@
+export interface HashPassword {
+  hash: (password: string) => Promise<string>
+}
+
 export interface PasswordValidator {
   validate: (password: string, hashedPassword: string) => Promise<boolean>
 }
